Make the api call inside HomePage
