+++ src/App.js
@@ -3,23 +3,7 @@ import { BrowserRouter as Router, Route } from 'react-router-dom'
 import { HomePage } from './routes/HomePage'
 
 class App extends React.Component {
-  constructor(props) {
-    super(props)
-
-    this.state = {
-      recipes: [],
-    }
-  }
-
-  async componentDidMount() {
-    const response = await fetch('/api/recipes/all.json')
-    const recipes = await response.json()
-    this.setState({ recipes })
-  }
-
   render() {
-    const { recipes } = this.state
-
     return (
       <>
         <nav className="d-flex justify-content-center">
